AVExplainer/requirements.txt
torch~=2.0.0+cu118
PyYAML~=6.0
opencv-python~=4.7.0.72
matplotlib~=3.7.1
numpy~=1.23.5
pandas~=1.5.3
seaborn~=0.12.2
Pillow~=9.5.0
scipy~=1.10.1
torchvision~=0.15.1+cu118
tqdm~=4.65.0
thop~=0.1.1.post2209072238
requests~=2.28.2
scikit-learn~=1.2.2
detectron2~=0.6
pycocotools~=2.0.6
iopath~=0.1.9
omegaconf~=2.3.0
fvcore~=0.1.5.post20221221
tabulate~=0.9.0
psutil~=5.9.4
setuptools~=58.1.0
termcolor~=2.2.0
packaging~=23.0
cloudpickle~=2.2.1